Remove dedicated Audi type, replace with eTron
