--- templates/definition/vehicle/audi-etron.yaml
@@ -1,14 +1,0 @@
-template: etron
-products:
-  - brand: Audi
-    description:
-      generic: e-tron
-params:
-  - preset: vehicle-base
-  - preset: vehicle-identify
-  - name: vin
-    example: WAUZZZ...
-render: |
-  type: etron
-  {{ include "vehicle-base" . }}
-  {{ include "vehicle-identify" . }}
