--- templates/docs/vehicle/etron_0.yaml
@@ -1,30 +1,0 @@
-product:
-  brand: Audi
-  description: e-tron
-render:
-  - default: |
-      type: template
-      template: etron
-      title: # Wird in der Benutzeroberfläche angezeigt (Optional)
-      user: # Benutzerkonto (bspw. E-Mail Adresse, User Id, etc.)
-      password: # Passwort des Benutzerkontos (bei führenden Nullen bitte in einfache Hochkommata setzen)
-      vin: WAUZZZ... # Erforderlich, wenn mehrere Fahrzeuge des Herstellers vorhanden sind (Optional)
-      capacity: 50 # Akkukapazität in kWh (Optional)
-    advanced: |
-      type: template
-      template: etron
-      title: # Wird in der Benutzeroberfläche angezeigt (Optional)
-      user: # Benutzerkonto (bspw. E-Mail Adresse, User Id, etc.)
-      password: # Passwort des Benutzerkontos (bei führenden Nullen bitte in einfache Hochkommata setzen)
-      vin: WAUZZZ... # Erforderlich, wenn mehrere Fahrzeuge des Herstellers vorhanden sind (Optional)
-      capacity: 50 # Akkukapazität in kWh (Optional)
-      phases: 3 # Die maximale Anzahl der Phasen welche genutzt werden können (Optional)
-      icon: car # Icon in der Benutzeroberfläche (Optional)
-      cache: 15m # Zeitintervall nach dem Daten erneut vom Fahrzeug abgefragt werden (Optional)
-      mode: # Möglich sind Off, Now, MinPV und PV, oder leer wenn keiner definiert werden soll (Optional)
-      minSoc: 25 # Ladung mit maximaler Geschwindigkeit bis zu dem angegeben Ladestand unabhängig PV-Erzeugung, wenn der Lademodus nicht auf 'Aus' steht (Optional)
-      targetSoc: 80 # Bis zu welchem Ladestand (Soc) soll das Fahrzeug geladen werden (Optional)
-      minCurrent: 6 # Definiert die minimale Stromstärke pro angeschlossener Phase mit welcher das Fahrzeug geladen werden soll (Optional)
-      maxCurrent: 16 # Definiert die maximale Stromstärke pro angeschlossener Phase mit welcher das Fahrzeug geladen werden soll (Optional)
-      identifiers: # Kann meist erst später eingetragen werden, siehe: https://docs.evcc.io/docs/guides/vehicles/#erkennung-des-fahrzeugs-an-der-wallbox (Optional)
-      priority: # Priorität des Ladepunktes oder Fahrzeugs in Relation zu anderen Ladepunkten oder Fahrzeugen für die Zuweisung von PV-Energie (Optional)
