+++ vehicle/audi.go
@@ -1,28 +1,31 @@
 package vehicle
 
 import (
+	"context"
 	"time"
 
 	"github.com/evcc-io/evcc/api"
 	"github.com/evcc-io/evcc/util"
 	"github.com/evcc-io/evcc/util/request"
-	"github.com/evcc-io/evcc/vehicle/audi"
+	"github.com/evcc-io/evcc/vehicle/audi/etron"
 	"github.com/evcc-io/evcc/vehicle/vag/idkproxy"
 	"github.com/evcc-io/evcc/vehicle/vag/service"
-	"github.com/evcc-io/evcc/vehicle/vw"
+	"github.com/evcc-io/evcc/vehicle/vag/vwidentity"
+	"github.com/evcc-io/evcc/vehicle/vw/id"
 )
 
-// https://github.com/davidgiga1993/AudiAPI
 // https://github.com/TA2k/ioBroker.vw-connect
+// https://github.com/arjenvrh/audi_connect_ha/blob/master/custom_components/audiconnect/audi_services.py
 
 // Audi is an api.Vehicle implementation for Audi cars
 type Audi struct {
 	*embed
-	*vw.Provider // provides the api implementations
+	*id.Provider // provides the api implementations
 }
 
 func init() {
 	registry.Add("audi", NewAudiFromConfig)
+	registry.Add("etron", NewAudiFromConfig)
 }
 
 // NewAudiFromConfig creates a new vehicle
@@ -51,21 +54,42 @@ func NewAudiFromConfig(other map[string]interface{}) (api.Vehicle, error) {
 
 	log := util.NewLogger("audi").Redact(cc.User, cc.Password, cc.VIN)
 
-	idk := idkproxy.New(log, audi.IDKParams)
-	ts, err := service.MbbTokenSource(log, idk, audi.AuthClientID, audi.AuthParams, cc.User, cc.Password)
+	// get initial VW identity id_token
+	q, err := vwidentity.Login(log, etron.AuthParams, cc.User, cc.Password)
 	if err != nil {
 		return nil, err
 	}
 
-	api := vw.NewAPI(log, ts, audi.Brand, audi.Country)
-	api.Client.Timeout = cc.Timeout
+	// exchange initial VW identity id_token for Audi AAZS token
+	idk := idkproxy.New(log, etron.IDKParams)
+	ats, its, err := service.AAZSTokenSource(log, idk, etron.AZSConfig, q)
+	if err != nil {
+		return nil, err
+	}
 
-	cc.VIN, err = ensureVehicle(cc.VIN, api.Vehicles)
+	// use the etron API for list of vehicles
+	api := etron.NewAPI(log, ats)
+
+	vehicle, err := ensureVehicleEx(
+		cc.VIN, func() ([]etron.Vehicle, error) {
+			ctx, cancel := context.WithTimeout(context.Background(), cc.Timeout)
+			defer cancel()
+			return api.Vehicles(ctx)
+		},
+		func(v etron.Vehicle) string {
+			return v.VIN
+		},
+	)
 
 	if err == nil {
-		if err = api.HomeRegion(cc.VIN); err == nil {
-			v.Provider = vw.NewProvider(api, cc.VIN, cc.Cache)
+		if v.Title_ == "" {
+			v.Title_ = vehicle.Nickname
 		}
+
+		api := id.NewAPI(log, its)
+		api.Client.Timeout = cc.Timeout
+
+		v.Provider = id.NewProvider(api, vehicle.VIN, cc.Cache)
 	}
 
 	return v, err
